UNet-MobileNet-Pytorch/requirements.txt
matplotlib
numpy
Pillow
torch
torchvision
torchsummary
tensorboard
future
tqdm
uvicorn
fastapi
python-multipart
httpx
loguru